Fix Fabric exception after manual verification
For obtaining the login connections, Fabric relies on the client's
current screen being a ConnectScreen. If it receives a login-related
network while a screen other than a ConnectScreen is active, an
exception is thrown in certain cases. Therefore, it is important for
AutoModpack to return to the ConnectScreen once it is done.