Implement certificate chain verification using standard library classes (#352)
* Implement certificate chain verification using standard library classes

* Fix IDE warnings

* Fix compilation failure caused by merge conflict

* Remove duplicate insertion of new known host

* Remove fingerprint from DataPacket

* Save known hosts after manual validation

* Apply review comments

* Make connection errors less verbose on client and on server side

* Extract first connection of DownloadClient from loop

* Fix Fabric exception after manual verification

For obtaining the login connections, Fabric relies on the client's
current screen being a ConnectScreen. If it receives a login-related
network while a screen other than a ConnectScreen is active, an
exception is thrown in certain cases. Therefore, it is important for
AutoModpack to return to the ConnectScreen once it is done.